Remove superfluous float division casting in SRM

+++ toolkit/srm/srm.py
@@ -323,7 +323,7 @@ class SRM(BaseEstimator):
             # M-step
 
             # Update Sigma_s and compute its trace
-            sigma_s = inv_sigma_s_rhos + shared_response.dot(shared_response.T) / float(samples)
+            sigma_s = inv_sigma_s_rhos + shared_response.dot(shared_response.T) / samples
             trace_sigma_s = samples * np.trace(sigma_s)
 
             # Update each subject's mapping transform W_i and error variance rho_i^2
@@ -338,7 +338,7 @@ class SRM(BaseEstimator):
                 rho2[subject] = trace_xtx[subject]
                 rho2[subject] += -2 * np.sum(w[subject] * a_subject).sum()
                 rho2[subject] += trace_sigma_s
-                rho2[subject] /= float(samples * voxels[subject])
+                rho2[subject] /= samples * voxels[subject]
 
             if self.verbose:
                 # Calculate and print the current log-likelihood for checking convergence
